pipelines/iot_analytics/requirements.txt
apache-beam[gcp]==2.61.0
pandas
scikit-learn
## Below dependencies are required if you have to run script /
google-cloud-bigtable
pandas
google-cloud-pubsub
tabulate